Revert "commit"
This reverts commit 4f6d7f15d664121e27fe9e675daf622890a55722.

+++ java/src/som/zd/sort/InsertSort.java
@@ -2,11 +2,11 @@ package som.zd.sort;
 
 public class InsertSort {
 	int n=5;
-	long[] a={1,2,8,9,5};
+	int[] a={1,2,8,9,5};
 	public void insertSort(){
 		int out,in;
 		for(out=1;out<n;out++){
-			long temp = a[out];
+			int temp = a[out];
 			in = out;
 			while(in>0 && a[in-1]>=temp){
 				a[in] = a[in-1];
